Update swagger-java-client and fix relevant tests
