+++ dockstore-webservice/src/main/java/io/swagger/model/ToolType.java
@@ -1,24 +1,7 @@
-/*
- *    Copyright 2016 OICR
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
 package io.swagger.model;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
@@ -29,7 +12,7 @@ import io.swagger.annotations.ApiModelProperty;
  **/
 
 @ApiModel(description = "Describes a type of tool allowing us to categorize workflows, the language of the workflow, tools, and maybe even other entities separately")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JaxRSServerCodegen", date = "2016-01-29T22:00:17.650Z")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2016-03-11T20:14:17.098Z")
 public class ToolType   {
   
   private String id = null;
@@ -40,6 +23,11 @@ public class ToolType   {
   /**
    * The unique identifier for the type
    **/
+  public ToolType id(String id) {
+    this.id = id;
+    return this;
+  }
+
   
   @ApiModelProperty(value = "The unique identifier for the type")
   @JsonProperty("id")
@@ -54,6 +42,11 @@ public class ToolType   {
   /**
    * A short friendly name for the type
    **/
+  public ToolType name(String name) {
+    this.name = name;
+    return this;
+  }
+
   
   @ApiModelProperty(value = "A short friendly name for the type")
   @JsonProperty("name")
@@ -68,6 +61,11 @@ public class ToolType   {
   /**
    * A longer explanation of what this type is and what it can accomplish
    **/
+  public ToolType description(String description) {
+    this.description = description;
+    return this;
+  }
+
   
   @ApiModelProperty(value = "A longer explanation of what this type is and what it can accomplish")
   @JsonProperty("description")
