+++ swagger-java-client/src/main/java/io/swagger/client/api/ContainertagsApi.java
@@ -11,7 +11,7 @@ import java.util.*;
 
 import java.util.*;
 
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class ContainertagsApi {
   private ApiClient apiClient;
 
@@ -35,7 +35,7 @@ public class ContainertagsApi {
   /**
    * Get tags  for a container by id
    * Lists tags for a container. Enter full path (include quay.io in path).
-   * @param containerId Container to modify.
+   * @param containerId Tool to modify.
    * @return List<Tag>
    */
   public List<Tag> getTagsByPath (Long containerId) throws ApiException {
@@ -82,7 +82,7 @@ public class ContainertagsApi {
   /**
    * Update the tags linked to a container
    * Tag correspond to each row of the versions table listing all information for a docker repo tag
-   * @param containerId Container to modify.
+   * @param containerId Tool to modify.
    * @param body List of modified tags
    * @return List<Tag>
    */
@@ -135,7 +135,7 @@ public class ContainertagsApi {
   /**
    * Add new tags linked to a container
    * Tag correspond to each row of the versions table listing all information for a docker repo tag
-   * @param containerId Container to modify.
+   * @param containerId Tool to modify.
    * @param body List of new tags
    * @return List<Tag>
    */
@@ -188,7 +188,7 @@ public class ContainertagsApi {
   /**
    * Delete tag linked to a container
    * Tag correspond to each row of the versions table listing all information for a docker repo tag
-   * @param containerId Container to modify.
+   * @param containerId Tool to modify.
    * @param tagId Tag to delete
    * @return void
    */
