+++ swagger-java-client/src/main/java/io/swagger/client/api/IntegrationbitbucketorgApi.java
@@ -10,7 +10,7 @@ import io.swagger.client.model.BitbucketOrgView;
 
 import java.util.*;
 
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class IntegrationbitbucketorgApi {
   private ApiClient apiClient;
