+++ swagger-java-client/src/main/java/io/swagger/client/model/Group.java
@@ -14,7 +14,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This describes a grouping of end-users for the purposes of managing sharing. Implementation-specific.
  **/
 @ApiModel(description = "This describes a grouping of end-users for the purposes of managing sharing. Implementation-specific.")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class Group   {
   
   private Long id = null;
