+++ swagger-java-client/src/main/java/io/swagger/client/model/Label.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * This describes a descriptive label that can be placed on an entry in the dockstore
  **/
 @ApiModel(description = "This describes a descriptive label that can be placed on an entry in the dockstore")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class Label   {
   
   private Long id = null;
