+++ swagger-java-client/src/main/java/io/swagger/client/model/ToolDescriptor.java
@@ -12,10 +12,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * A tool descriptor is a metadata document that describes one or more tools.
  **/
 @ApiModel(description = "A tool descriptor is a metadata document that describes one or more tools.")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class ToolDescriptor   {
   
   private String descriptor = null;
+  private String url = null;
 
   
   /**
@@ -31,6 +32,19 @@ public class ToolDescriptor   {
   }
 
   
+  /**
+   * Optional url to the tool descriptor used to build this image, should include version information, and can include a git hash (e.g. https://raw.githubusercontent.com/ICGC-TCGA-PanCancer/pcawg_delly_workflow/ea2a5db69bd20a42976838790bc29294df3af02b/delly_docker/Delly.cwl )
+   **/
+  @ApiModelProperty(value = "Optional url to the tool descriptor used to build this image, should include version information, and can include a git hash (e.g. https://raw.githubusercontent.com/ICGC-TCGA-PanCancer/pcawg_delly_workflow/ea2a5db69bd20a42976838790bc29294df3af02b/delly_docker/Delly.cwl )")
+  @JsonProperty("url")
+  public String getUrl() {
+    return url;
+  }
+  public void setUrl(String url) {
+    this.url = url;
+  }
+
+  
 
   @Override
   public String toString()  {
@@ -38,6 +52,7 @@ public class ToolDescriptor   {
     sb.append("class ToolDescriptor {\n");
     
     sb.append("    descriptor: ").append(StringUtil.toIndentedString(descriptor)).append("\n");
+    sb.append("    url: ").append(StringUtil.toIndentedString(url)).append("\n");
     sb.append("}");
     return sb.toString();
   }
