+++ swagger-java-client/src/main/java/io/swagger/client/model/ToolDockerfile.java
@@ -12,10 +12,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * A tool dockerfile is a document that describes how to build a particular Docker image.
  **/
 @ApiModel(description = "A tool dockerfile is a document that describes how to build a particular Docker image.")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class ToolDockerfile   {
   
   private String dockerfile = null;
+  private String url = null;
 
   
   /**
@@ -31,6 +32,19 @@ public class ToolDockerfile   {
   }
 
   
+  /**
+   * Optional url to the dockerfile used to build this image, should include version information, and can include a git hash  (e.g. https://raw.githubusercontent.com/ICGC-TCGA-PanCancer/pcawg_delly_workflow/c83478829802b4d36374870843821abe1b625a71/delly_docker/Dockerfile )
+   **/
+  @ApiModelProperty(value = "Optional url to the dockerfile used to build this image, should include version information, and can include a git hash  (e.g. https://raw.githubusercontent.com/ICGC-TCGA-PanCancer/pcawg_delly_workflow/c83478829802b4d36374870843821abe1b625a71/delly_docker/Dockerfile )")
+  @JsonProperty("url")
+  public String getUrl() {
+    return url;
+  }
+  public void setUrl(String url) {
+    this.url = url;
+  }
+
+  
 
   @Override
   public String toString()  {
@@ -38,6 +52,7 @@ public class ToolDockerfile   {
     sb.append("class ToolDockerfile {\n");
     
     sb.append("    dockerfile: ").append(StringUtil.toIndentedString(dockerfile)).append("\n");
+    sb.append("    url: ").append(StringUtil.toIndentedString(url)).append("\n");
     sb.append("}");
     return sb.toString();
   }
