+++ swagger-java-client/src/main/java/io/swagger/client/model/ToolType.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * Describes a type of tool allowing us to categorize workflows, the language of the workflow, tools, and maybe even other entities separately
  **/
 @ApiModel(description = "Describes a type of tool allowing us to categorize workflows, the language of the workflow, tools, and maybe even other entities separately")
-@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-01T15:18:10.919-05:00")
+@javax.annotation.Generated(value = "class io.swagger.codegen.languages.JavaClientCodegen", date = "2016-03-11T15:28:43.725-05:00")
 public class ToolType   {
   
   private String id = null;
